homework8

See merge request nikitalisenkov/tfs_spring_2024!8

+++ app/src/main/java/com/example/app/data/network/model/StreamResponse.kt
@@ -9,7 +9,7 @@ data class StreamsResponse(
 
 data class StreamResponse(
     @SerializedName("stream_id")
-    val streamId: Int,
+    val streamId: Long,
     @SerializedName("name")
     val name: String,
     @SerializedName("color")
